+++ src/main/java/net/soulsweaponry/enchantments/FrostsilverEnchantment.java
@@ -1,0 +1,29 @@
+package net.soulsweaponry.enchantments;
+
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.enchantment.EnchantmentTarget;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.item.ItemStack;
+import net.soulsweaponry.items.gun.GunItem;
+
+public class FrostsilverEnchantment extends Enchantment {
+
+    public FrostsilverEnchantment(Rarity weight, EquipmentSlot... slotTypes) {
+        super(weight, EnchantmentTarget.BREAKABLE, slotTypes);
+    }
+
+    @Override
+    public int getMinPower(int level) {
+        return 10 + level * 10;
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 2;
+    }
+
+    @Override
+    public boolean isAcceptableItem(ItemStack stack) {
+        return stack.getItem() instanceof GunItem;
+    }
+}