+++ src/main/java/net/soulsweaponry/items/gun/Blunderbuss.java
@@ -5,17 +5,11 @@ import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.ItemStack;
-import net.minecraft.particle.ParticleTypes;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvents;
-import net.minecraft.stat.Stats;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
-import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
 import net.soulsweaponry.registry.EnchantRegistry;
-import net.soulsweaponry.registry.ItemRegistry;
 
 public class Blunderbuss extends GunItem {
 
@@ -50,7 +44,7 @@ public class Blunderbuss extends GunItem {
     }
 
     @Override
-    public int bulletsNeeded() {
+    public int getBulletsNeeded() {
         return ConfigConstructor.blunderbuss_bullets_needed;
     }
 
@@ -66,36 +60,15 @@ public class Blunderbuss extends GunItem {
             this.notifyDisabled(user);
             return TypedActionResult.fail(user.getStackInHand(hand));
         }
-        boolean bl = user.getAbilities().creativeMode || EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0;
-        ItemStack itemStack = user.getProjectileType(stack);
-        if (!itemStack.isEmpty() || bl) {
-            if (itemStack.isEmpty()) {
-                itemStack = new ItemStack(ItemRegistry.SILVER_BULLET);
-            }
-            boolean bl2 = bl && itemStack.isOf(ItemRegistry.SILVER_BULLET);
+        ItemStack itemStack = this.canShoot(user, stack);
+        if (itemStack != null) {
             int projectileCount = ConfigConstructor.blunderbuss_projectile_amount + EnchantmentHelper.getLevel(Enchantments.POWER, stack) / 2;
-            Vec3d pov = user.getRotationVector();
-            Vec3d particleBox = pov.multiply(1).add(user.getPos());
             for (int i = 0; i < projectileCount; i++) {
                 PersistentProjectileEntity entity = this.createSilverBulletEntity(world, user, stack);
                 world.spawnEntity(entity);
             }
-            if (world.isClient) {
-                for (int k = 0; k < 50; k++) {
-                    world.addParticle(ParticleTypes.FLAME, true, particleBox.x, particleBox.y + 1.5F, particleBox.z, pov.x + user.getRandom().nextDouble() - .5, pov.y + user.getRandom().nextDouble() - .5, pov.z + user.getRandom().nextDouble() - .5);
-                    world.addParticle(ParticleTypes.SMOKE, true, particleBox.x, particleBox.y + 1.5F, particleBox.z, pov.x + user.getRandom().nextDouble() - .5, pov.y + user.getRandom().nextDouble() - .5, pov.z + user.getRandom().nextDouble() - .5);
-                }
-            }
-            world.playSound(user, user.getBlockPos(), SoundEvents.ENTITY_GENERIC_EXPLODE, SoundCategory.PLAYERS, 1f, 1f);
-            stack.damage(1, user, (p_220045_0_) -> p_220045_0_.sendToolBreakStatus(user.getActiveHand()));
-            if (!bl2 && !user.getAbilities().creativeMode) {
-                itemStack.decrement(this.bulletsNeeded());
-                if (itemStack.isEmpty()) {
-                    user.getInventory().removeOne(itemStack);
-                }
-            }
-            user.incrementStat(Stats.USED.getOrCreateStat(this));
-            if (!user.isCreative()) user.getItemCooldownManager().set(this, this.getCooldown(stack));
+            this.spawnShotParticles(world, user, 50, 0.2f);
+            this.postShot(world, user, stack);
             return TypedActionResult.success(stack, world.isClient());
         }
         return TypedActionResult.fail(stack);