+++ src/main/java/net/soulsweaponry/items/gun/HunterCannon.java
@@ -6,20 +6,14 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.ItemStack;
-import net.minecraft.particle.ParticleTypes;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvents;
-import net.minecraft.stat.Stats;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
 import net.soulsweaponry.entity.projectile.Cannonball;
 import net.soulsweaponry.entity.projectile.SilverBulletEntity;
 import net.soulsweaponry.registry.EnchantRegistry;
-import net.soulsweaponry.registry.ItemRegistry;
+import net.soulsweaponry.util.WeaponUtil;
 
 public class HunterCannon extends GunItem {
 
@@ -54,7 +48,7 @@ public class HunterCannon extends GunItem {
     }
 
     @Override
-    public int bulletsNeeded() {
+    public int getBulletsNeeded() {
         return ConfigConstructor.hunter_cannon_bullets_needed;
     }
 
@@ -70,49 +64,21 @@ public class HunterCannon extends GunItem {
             return TypedActionResult.fail(user.getStackInHand(hand));
         }
         ItemStack stack = user.getStackInHand(hand);
-        boolean bl = user.getAbilities().creativeMode || EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0;
-        ItemStack itemStack = user.getProjectileType(stack);
-        int bulletsNeeded = this.bulletsNeeded();
-        if (!itemStack.isEmpty() && itemStack.getCount() >= bulletsNeeded || bl) {
-            if (itemStack.isEmpty()) {
-                itemStack = new ItemStack(ItemRegistry.SILVER_BULLET);
-            }
-            boolean bl2 = bl && itemStack.isOf(ItemRegistry.SILVER_BULLET);
-            Vec3d pov = user.getRotationVector();
-            Vec3d particleBox = pov.multiply(1).add(user.getPos());
-            if (world.isClient) {
-                for (int k = 0; k < 50; k++) {
-                    world.addParticle(ParticleTypes.FLAME, true, particleBox.x, particleBox.y + 1.5F, particleBox.z, pov.x + user.getRandom().nextDouble() - .5, pov.y + user.getRandom().nextDouble() - .5, pov.z + user.getRandom().nextDouble() - .5);
-                    world.addParticle(ParticleTypes.SMOKE, true, particleBox.x, particleBox.y + 1.5F, particleBox.z, pov.x + user.getRandom().nextDouble() - .5, pov.y + user.getRandom().nextDouble() - .5, pov.z + user.getRandom().nextDouble() - .5);
-                }
-            }
-
+        ItemStack itemStack = this.canShoot(user, stack);
+        if (itemStack != null) {
+            this.spawnShotParticles(world, user, 50, 0.4f);
             PersistentProjectileEntity entity = this.createSilverBulletEntity(world, user, stack);
             world.spawnEntity(entity);
-            world.playSound(user, user.getBlockPos(), SoundEvents.ENTITY_GENERIC_EXPLODE, SoundCategory.PLAYERS, 1f, 1f);
-            stack.damage(bulletsNeeded, user, (p_220045_0_) -> p_220045_0_.sendToolBreakStatus(user.getActiveHand()));
-            if (!bl2 && !user.getAbilities().creativeMode) {
-                itemStack.decrement(bulletsNeeded);
-                if (itemStack.isEmpty()) {
-                    user.getInventory().removeOne(itemStack);
-                }
-            }
-            float f = user.getYaw();
-            float g = user.getPitch();
-            float h = -MathHelper.sin(f * 0.017453292F) * MathHelper.cos(g * 0.017453292F);
-            float k = -MathHelper.sin(g * 0.017453292F);
-            float l = MathHelper.cos(f * 0.017453292F) * MathHelper.cos(g * 0.017453292F);
-            float m = MathHelper.sqrt(h * h + k * k + l * l);
-            float n = 2.0F * ((1.0F + 1F) / 4.0F);
-            h *= n / m;
-            k *= n / m;
-            l *= n / m;
-            user.addVelocity(-(double)h, -(double)k, -(double)l);
-            user.incrementStat(Stats.USED.getOrCreateStat(this));
-            if (!user.isCreative()) user.getItemCooldownManager().set(this, this.getCooldown(stack));
+            WeaponUtil.launchTarget(user, 2f, true);
+            this.postShot(world, user, stack);
             return TypedActionResult.success(stack, world.isClient());
         }
         return TypedActionResult.fail(stack);
+    }
+
+    @Override
+    public int getStackDamageToApply() {
+        return this.getBulletsNeeded();
     }
 
     @Override