+++ src/main/java/net/soulsweaponry/items/gun/HunterPistol.java
@@ -1,21 +1,14 @@
 package net.soulsweaponry.items.gun;
 
 import net.minecraft.enchantment.EnchantmentHelper;
-import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.ItemStack;
-import net.minecraft.particle.ParticleTypes;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvents;
-import net.minecraft.stat.Stats;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
-import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
 import net.soulsweaponry.registry.EnchantRegistry;
-import net.soulsweaponry.registry.ItemRegistry;
 
 public class HunterPistol extends GunItem {
 
@@ -50,7 +43,7 @@ public class HunterPistol extends GunItem {
     }
 
     @Override
-    public int bulletsNeeded() {
+    public int getBulletsNeeded() {
         return ConfigConstructor.hunter_pistol_bullets_needed;
     }
 
@@ -66,35 +59,12 @@ public class HunterPistol extends GunItem {
             return TypedActionResult.fail(user.getStackInHand(hand));
         }
         ItemStack stack = user.getStackInHand(hand);
-        boolean bl = user.getAbilities().creativeMode || EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0;
-        ItemStack itemStack = user.getProjectileType(stack);
-        if (!itemStack.isEmpty() || bl) {
-            if (itemStack.isEmpty()) {
-                itemStack = new ItemStack(ItemRegistry.SILVER_BULLET);
-            }
-            boolean bl2 = bl && itemStack.isOf(ItemRegistry.SILVER_BULLET);
-            Vec3d pov = user.getRotationVector();
-            Vec3d particleBox = pov.multiply(1).add(user.getPos());
-            if (world.isClient) {
-                for (int k = 0; k < 10; k++) {
-                    world.addParticle(ParticleTypes.FLAME, true, particleBox.x, particleBox.y + 1.5F, particleBox.z, pov.x + user.getRandom().nextDouble() - .25, pov.y + user.getRandom().nextDouble() - .5, pov.z + user.getRandom().nextDouble() - .25);
-                    world.addParticle(ParticleTypes.SMOKE, true, particleBox.x, particleBox.y + 1.5F, particleBox.z, pov.x + user.getRandom().nextDouble() - .25, pov.y + user.getRandom().nextDouble() - .5, pov.z + user.getRandom().nextDouble() - .25);
-                }
-            }
-
+        ItemStack itemStack = this.canShoot(user, stack);
+        if (itemStack != null) {
+            this.spawnShotParticles(world, user, 10, 0.1f);
             PersistentProjectileEntity entity = this.createSilverBulletEntity(world, user, stack);
             world.spawnEntity(entity);
-            world.playSound(user, user.getBlockPos(), SoundEvents.ENTITY_GENERIC_EXPLODE, SoundCategory.PLAYERS, 1f,1f);
-            stack.damage(1, user, (p_220045_0_) -> p_220045_0_.sendToolBreakStatus(user.getActiveHand()));
-            if (!bl2 && !user.getAbilities().creativeMode) {
-                itemStack.decrement(this.bulletsNeeded());
-                if (itemStack.isEmpty()) {
-                    user.getInventory().removeOne(itemStack);
-                }
-            }
-
-            user.incrementStat(Stats.USED.getOrCreateStat(this));
-            if (!user.isCreative()) user.getItemCooldownManager().set(this, this.getCooldown(stack));
+            this.postShot(world, user, stack);
             return TypedActionResult.success(stack, world.isClient());
         }
         return TypedActionResult.fail(stack);